requirement.txt
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.8.0

